bitkeeper revision 1.1159.1.565 (420b7bafUdB4uOmG-Qm1AYMify0jrQ)
p2m and m2p tables are now 4 bytes per entry always, even on x86/64.
Fixed mapping of m2p table into kernel space on x86/64.

+++ xen/include/public/arch-x86_32.h
@@ -71,7 +71,7 @@
  */
 #define HYPERVISOR_VIRT_START (0xFC000000UL)
 #ifndef machine_to_phys_mapping
-#define machine_to_phys_mapping ((unsigned long *)HYPERVISOR_VIRT_START)
+#define machine_to_phys_mapping ((u32 *)HYPERVISOR_VIRT_START)
 #endif
 
 #ifndef __ASSEMBLY__
